Rename secondary nav component to sub nav
Even though we're using it for secondary navigation the actual component
is called sub navigation. This change will keep things aligned and open
the possibility of us also using the secondary (tab style) nav at some
point

https://x-govuk.github.io/govuk-prototype-components/sub-navigation/
