+++ app/views/layouts/admin.html.erb
@@ -23,7 +23,12 @@
 
         <div class="govuk-grid-row">
           <div class="govuk-grid-column-one-quarter">
-            <%= render NpqSeparation::Admin::SecondaryNavigationComponent.new(request.path) %>
+            <%=
+              render NpqSeparation::Admin::SubNavigationComponent.new(
+                request.path,
+                lead_providers: LeadProviders::Find.new.all
+              )
+            %>
           </div>
 
           <div class="govuk-grid-column-three-quarters">
