+++ spec/components/npq_separation/admin/sub_navigation_component_spec.rb
@@ -1,12 +1,12 @@
 require "rails_helper"
 
-RSpec.describe NpqSeparation::Admin::SecondaryNavigationComponent, type: :component do
+RSpec.describe NpqSeparation::Admin::SubNavigationComponent, type: :component do
   let(:fake_logger) { object_double(Rails.logger, error: true) }
   let(:path) { "/npq-separation/admin" }
 
   before do
     allow(Rails).to receive(:logger).and_return(fake_logger)
-    render_inline(NpqSeparation::Admin::SecondaryNavigationComponent.new(path))
+    render_inline(NpqSeparation::Admin::SubNavigationComponent.new(path))
   end
 
   it "renders a visually hidden level 2 heading" do
